Add Doxygen docs and separate .cpp files for C++ NeoPixel transport
